funfa

+++ main.c
@@ -6,19 +6,20 @@
 /*   By: mde-figu <mde-figu@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/05 23:40:16 by mde-figu          #+#    #+#             */
-/*   Updated: 2021/07/06 00:02:18 by mde-figu         ###   ########.fr       */
+/*   Updated: 2021/07/06 12:52:21 by mde-figu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include <stdint.h>
 
-int    main(int x)
+int    main(uint64_t x)
 {
-const uint64_t m1  = 0x5555555555555555; //binary: 0101...
-const uint64_t m2  = 0x3333333333333333; //binary: 00110011..
-const uint64_t m4  = 0x0f0f0f0f0f0f0f0f; //binary:  4 zeros,  4 ones ...
-const uint64_t h01 = 0x0101010101010101; //the sum of 256 to the power of 0,1,2,3...
+
+    const uint64_t m1  = 0x5555555555555555; //binary: 0101...
+    const uint64_t m2  = 0x3333333333333333; //binary: 00110011..
+    const uint64_t m4  = 0x0f0f0f0f0f0f0f0f; //binary:  4 zeros,  4 ones ...
+    const uint64_t h01 = 0x0101010101010101; //the sum of 256 to the power of 0,1,2,3...
 
     x -= (x >> 1) & m1;             //put count of each 2 bits into those 2 bits
     x = (x & m2) + ((x >> 2) & m2); //put count of each 4 bits into those 4 bits
